20250607 Limpieza de funciones y restrucutración, valgrind+
