+++ include/cube3d.h
@@ -54,6 +54,7 @@ typedef struct s_game
 // Init
 void		init_game_window(t_game *game, char *map_path);
 void		init_map(t_map *map);
+int			map_setup(t_map *map, const char *map_name, char *full_path, size_t max_len);
 
 // Parse
 int			map_parse(t_map *map, char *map_path);
@@ -62,7 +63,6 @@ int			parse_texture_colors(t_map *map, char *line);
 
 // Validation
 int			map_validate(t_map *map);
-int			validate_map_path(const char *path);
 int			validate_walls(char **map, int start_x, int start_y);
 int			validate_colors_and_textures(t_map *map);
