+++ valgrind_out_def.txt
@@ -1,0 +1,10 @@
+==17024==    definitely lost: 66,096 bytes in 3 blocks
+==17024==    indirectly lost: 0 bytes in 0 blocks
+==17024==      possibly lost: 0 bytes in 0 blocks
+==17024==    still reachable: 374,450 bytes in 3,352 blocks
+==17024==         suppressed: 0 bytes in 0 blocks
+==17024== Rerun with --leak-check=full to see details of leaked memory
+==17024== 
+==17024== Use --track-origins=yes to see where uninitialised values come from
+==17024== For lists of detected and suppressed errors, rerun with: -s
+==17024== ERROR SUMMARY: 342015 errors from 1000 contexts (suppressed: 0 from 0)
